+++ apps/backend-functions/src/rescue.ts
@@ -4,7 +4,8 @@ import { sign } from 'jsonwebtoken';
 import * as admin from 'firebase-admin';
 import { Request, Response } from 'firebase-functions';
 import { getStorageBucketName } from './internals/firebase';
-import { sendRequest } from './utils';
+import { sendRequest } from '@blockframes/firebase-utils';
+
 
 const linkDuration = 60 * 60 * 5; // 5 hours in seconds = 60 seconds * 60 minutes * 5 = 18 000 seconds
