+++ libs/firebase-utils/src/index.ts
@@ -11,6 +11,8 @@ export * from './lib/types';
 export * from './lib/algolia/algolia';
 export * from './lib/commands';
 export * from './lib/anonymize';
-export * from './lib/types'
-export * from './lib/algolia/algolia'
-export * from './lib/commands'
+export * from './lib/types';
+export * from './lib/algolia/algolia';
+export * from './lib/commands';
+export * from './lib/jwplayer-api';
+export * from './lib/node-request';
